refactor(DatingSite): refactored project
- added and used lombok
- structured project files
