+++ src/main/java/org/spring/datingsite/invitation/enums/InvitationStateEnum.java
@@ -1,4 +1,4 @@
-package org.spring.datingsite.enums;
+package org.spring.datingsite.invitation.enums;
 
 public enum InvitationStateEnum {
     NONE,
